Refs #16. refactor: move violini model to its own module
